+++ powerbox/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.5.6"
+__version__ = "0.5.7"
 
 from .powerbox import PowerBox, LogNormalPowerBox
 from .tools import get_power, angular_average, angular_average_nd