+++ tests/test_power.py
@@ -65,8 +65,37 @@ def test_power2d():
 
 
 def test_power3d():
-    pb = PowerBox(50, dim=2, pk=lambda k: 1.0 * k ** -2., boxlength=1.0, b=1)
+    pb = PowerBox(50, dim=3, pk=lambda k: 1.0 * k ** -2., boxlength=1.0, b=1)
     p, k = get_power(pb.delta_x(), pb.boxlength, b=1)
 
     print(p / (1.0 * k ** -2.))
     assert np.allclose(p, 1.0 * k ** -2., rtol=2)
+
+
+def test_k_zero_ignore():
+    pb = PowerBox(50, dim=2, pk=lambda k: 1.0 * k ** -2., boxlength=1.0, b=1)
+
+    dx = pb.delta_x()
+    p1, k1 = get_power(dx, pb.boxlength, bin_ave=False)
+    p0, k0 = get_power(dx, pb.boxlength, ignore_zero_mode=True, bin_ave=False)
+
+    assert np.all(k1 == k0)
+
+    assert np.all(p1[1:] == p0[1:])
+
+    assert p1[0] != p0[0]
+
+
+def test_k_weights():
+    pb = PowerBox(50, dim=2, pk=lambda k: 1.0 * k ** -2., boxlength=1.0, b=1)
+
+    dx = pb.delta_x()
+
+    k_weights = np.ones_like(dx)
+    k_weights[:, 25] = 0
+
+    p1, k1 = get_power(dx, pb.boxlength, bin_ave=False)
+    p0, k0 = get_power(dx, pb.boxlength, bin_ave=False, k_weights= k_weights)
+
+    assert np.all(k1 == k0)
+    assert not np.allclose(p1, p0)